DelegateButton.jsx refactored to fire on user click
